Use ic-cdk v0.5.7
