+++ rust/exchange_rate/Cargo.toml
@@ -9,8 +9,8 @@ name = "exchange_rate"
 path = "src/main.rs"
 
 [dependencies]
-ic-cdk = "0.5.0"
-ic-cdk-macros = "0.5.0"
+ic-cdk = "0.5.7"
+ic-cdk-macros = "0.5.7"
 serde = "1.0.126"
 candid = "0.7.14"
 serde_json = "1.0.81"
